requirements.txt
# Core dependencies
kiteconnect==4.2.0
python-dotenv==1.0.0
requests==2.31.0

# Logging and utilities
loguru==0.7.2

# Data analysis (stubs for future)
pandas==2.1.3
numpy==1.24.3

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
